Implement .clear for BufferedDisplay.
There's some ugliness here, the display buffer code appears to be
written to be minimalistic, so lines aren't actually initialized to the
underlying screen dimensions until you actually start writing to it. Not
sure what to make of that. But at least now, calling .clear on the
BufferedDisplay ought to sync things up.