Force binary mode on Windows to prevent newline compatibility problems.
Instead of trying to deal with the horrid mess that is Windows newline
endings, let's just standardize on the Unix newline style and write it
out in binary on Windows, so that save files from either build will work
on the other build.  Save files aren't supposed to be user-editable
anyway, so it's not a problem to require extra effort for developers to
decode.