Revert to SupportsWeight + BlocksMove for falling.
Basically, walking on top of a creature that can't hold you up will
cause you to damage him, but you will also roll off. So the old model
actually made sense. The creature blocks moves, but simultaneously is
unable to hold your weight, and hence the injury and the throwing to the
side: it buckles under your weight.