Emergency save should not crash if Game not initialized yet!!
Should not choke on stale savefiles from incompatible versions and abort
during startup; rename them instead and keep going.

Default to GUI mode for Windows. Console backend won't even start
because we compile the Windows executable as a GUI application. :-D (We
could *probably* get it to work by explicitly calling the Windows API to
create a console, but why bother now that the GUI backend works??)

Get rid of exit pause in Windows builds: we don't need this anymore.